Handle localized view markdown templates in pegasus and the sync.
Return locale specific view template, if exists, in pegasus.

We want to make sure that the 'views' markdown exists as a sub-directory
so that it won't interfere with the existing `public` markdown from the
pegasus directory while we translate the views paths. It is messy.

This allows localized view templates to exist and be used on certain
pages.

This does add a markdown testing crowdin repository.